build: use std=gnu99 instead of c99
because the new version of chicken.h won't work with c99, but it works
with gnu99 so we'll try to use that as long as that works.

note that it's tricky to remove std entirely because the warnings are
coming from the for loops in chicken_private.h which are currently
installed independently from this egg.