build: embedding chipmunk 6.2.1
because it will hopefully make cross-compilation and embedding of
chipmunk easier. with a working cross-chicken, you should be able to do

  cross-chicken-install chickmunk

and you've got the egg, without the need to port chickmunk to
android/ios as well - the chicken cross toolchain might be able to do
this automatically for us.